+++ src/mjolnir/graphtilebuilder.cc
@@ -1006,7 +1006,6 @@ void GraphTileBuilder::UpdateTrafficSegments(const bool update_dir_edges) {
   }
 
   // Open file and truncate
-  std::stringstream in_mem;
   std::ofstream file(filename.c_str(), std::ios::out | std::ios::binary | std::ios::trunc);
   if (file.is_open()) {
     // Write a new header
@@ -1076,6 +1075,85 @@ void GraphTileBuilder::AddTurnLanes(const uint32_t idx, const std::string& str) 
   if (!str.empty()) {
     uint32_t offset = AddName(str);
     turnlanes_builder_.emplace_back(idx, offset);
+  }
+}
+
+// Add a predicted speed profile for a directed edge.
+void GraphTileBuilder::AddPredictedSpeed(const uint32_t idx, const std::vector<int16_t>& profile) {
+  // Create the index builder on the first profile added. Resize to equal the count of
+  // directed edges
+  if (speed_profile_index_builder_.size() == 0) {
+    speed_profile_index_builder_.resize(header_->directededgecount());
+  }
+
+  if (idx < header_->directededgecount()) {
+    speed_profile_index_builder_[idx] = speed_profile_builder_.size() / kCoefficientCount;
+
+    // Append the profile
+    if (profile.size() == kCoefficientCount) {
+      speed_profile_builder_.reserve(speed_profile_builder_.size() + kCoefficientCount);
+      speed_profile_builder_.insert(speed_profile_builder_.end(), profile.begin(), profile.end());
+    } else {
+      throw std::runtime_error("GraphTileBuilder AddPredictedSpeed profile is not correct size: " +
+                               std::to_string(profile.size()));
+    }
+  } else {
+    throw std::runtime_error("GraphTileBuilder AddPredictedSpeed index is out of bounds");
+  }
+}
+
+// Updates a tile with predictive speed data. Also updates directed edges with
+// free flow and constrained flow speeds and the predicted traffic flag. The
+// predicted traffic is written after turn lane data.
+void GraphTileBuilder::UpdatePredictedSpeeds(const std::vector<DirectedEdge>& directededges) {
+
+  // Get the name of the file
+  boost::filesystem::path filename =
+      tile_dir_ + filesystem::path_separator + GraphTile::FileSuffix(header_builder_.graphid());
+
+  // Make sure the directory exists on the system
+  if (!boost::filesystem::exists(filename.parent_path()))
+    boost::filesystem::create_directories(filename.parent_path());
+
+  // Open file and truncate
+  std::ofstream file(filename.c_str(), std::ios::out | std::ios::binary | std::ios::trunc);
+  if (file.is_open()) {
+    // Write a new header - add the offset to predicted speed data and the profile count.
+    // Update the end offset (shift by the amount of predicted speed data added).
+    header_builder_.set_end_offset(header_->end_offset() +
+                                   (speed_profile_index_builder_.size() * sizeof(uint32_t)) +
+                                   (speed_profile_builder_.size() * sizeof(int16_t)));
+    size_t offset = header_->turnlane_offset() + header_->turnlane_count() * sizeof(TurnLanes);
+    header_builder_.set_predictedspeeds_offset(offset);
+    header_builder_.set_predictedspeeds_count(speed_profile_builder_.size() / kCoefficientCount);
+    file.write(reinterpret_cast<const char*>(&header_builder_), sizeof(GraphTileHeader));
+
+    // Copy the nodes (they are unchanged when adding predicted speeds).
+    file.write(reinterpret_cast<const char*>(nodes_), header_->nodecount() * sizeof(NodeInfo));
+
+    // Write the updated directed edges. Make sure edge count matches.
+    if (directededges.size() != header_->directededgecount()) {
+      throw std::runtime_error("GraphTileBuilder::Update - directed edge count has changed");
+    }
+    file.write(reinterpret_cast<const char*>(directededges.data()),
+               directededges.size() * sizeof(DirectedEdge));
+
+    // Write out data from access restrictions to the end of turn lane data
+    auto begin = reinterpret_cast<const char*>(&access_restrictions_[0]);
+    auto end = reinterpret_cast<const char*>(header()) + offset;
+    file.write(begin, end - begin);
+
+    // Append the speed profile indexes and profiles.
+    file.write(reinterpret_cast<const char*>(speed_profile_index_builder_.data()),
+               speed_profile_index_builder_.size() * sizeof(uint32_t));
+    file.write(reinterpret_cast<const char*>(speed_profile_builder_.data()),
+               speed_profile_builder_.size() * sizeof(int16_t));
+
+    // Write the rest of the tiles. TBD (if anything is added after the speed profiles
+    // then this will need to be updated)
+
+    // Close the file
+    file.close();
   }
 }